Revert "Moved conversion of token to userid and all logic out of adminQuizNameUpdate server route in server.ts"
This reverts commit 0b3b5f28bf186b14391125832bbdd7caa97a7cd5.

+++ src/helper.ts
@@ -2,7 +2,7 @@ import { getData } from './dataStore';
 import { User, Quiz } from './interface';
 
 export function getUserIdFromToken(sessionId: string): number {
-  const result = parseInt(sessionId);
+  const result = parseFloat(sessionId);
   const store = getData();
 
   const sessArr = store.sessions;
